+++ Source/ModLoader.sln
@@ -30,7 +30,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FluidWarpMod", "FluidWarpMo
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RoomSize", "RoomSizeMod\RoomSize.csproj", "{7D13E2D1-D400-4A50-8AFA-147EAF5375C1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NoFixedTempMod", "NoFixedTempMod\NoFixedTempMod.csproj", "{510A0A56-2D15-49B8-9932-BA232DC1F231}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NoFixedTemp", "NoFixedTemp\NoFixedTemp.csproj", "{510A0A56-2D15-49B8-9932-BA232DC1F231}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FluidPhysicsMod", "FluidPhysicsMod\FluidPhysicsMod.csproj", "{F4D08C87-B8C2-481F-9728-CE0D342EC2CE}"
 EndProject
@@ -304,7 +304,7 @@ Global
 		{DB33F126-AECD-4E02-9F1F-1E4B08DC27C4} = {8E4312B0-3AE8-42CB-A2ED-CB8E39B39CD4}
 		{74B7B353-BB3E-4EE1-9035-7357D5472D39} = {3DDC5DF0-0321-41EA-B556-022CEE509815}
 		{7D13E2D1-D400-4A50-8AFA-147EAF5375C1} = {3DDC5DF0-0321-41EA-B556-022CEE509815}
-		{510A0A56-2D15-49B8-9932-BA232DC1F231} = {F23F1C19-7F80-41CC-86F7-F720C6D3C43B}
+		{510A0A56-2D15-49B8-9932-BA232DC1F231} = {8E4312B0-3AE8-42CB-A2ED-CB8E39B39CD4}
 		{F4D08C87-B8C2-481F-9728-CE0D342EC2CE} = {3DDC5DF0-0321-41EA-B556-022CEE509815}
 		{4BF8EC3F-EB0C-4FC7-B2F5-458E91F63E96} = {F23F1C19-7F80-41CC-86F7-F720C6D3C43B}
 		{AE64BD07-9B27-4E16-9767-DD7894F70806} = {F23F1C19-7F80-41CC-86F7-F720C6D3C43B}
